MYST3: Split detection features & adapt to new plugins.
